Move revison/version column into Metadata namespace
